Remove benchmarks for task
The real benchmarks will be the compilation benchmarks, since they will
tell us the real overhead for a medium-sized and CPU-bounded scenario
(compilation of dependent projects).

Benchmarking tasks for trivial operations is not bringing us much, so we
remove this code with hope that the compilation benchmarks are the one
that shed light on bloop + monix performance.